+++ apps/builder/src/features/settings/components/TypingEmulationForm.tsx
@@ -42,14 +42,14 @@ export const TypingEmulationForm = ({ typingEmulation, onUpdate }: Props) => {
   return (
     <Stack spacing={6}>
       <SwitchWithRelatedSettings
-        label={'Typing emulation'}
+        label={'Emulação de escrita'}
         initialValue={
           typingEmulation?.enabled ?? defaultSettings.typingEmulation.enabled
         }
         onCheckChange={updateIsEnabled}
       >
         <NumberInput
-          label="Words per minutes:"
+          label="Palavras por minuto:"
           data-testid="speed"
           defaultValue={
             typingEmulation?.speed ?? defaultSettings.typingEmulation.speed
@@ -62,7 +62,7 @@ export const TypingEmulationForm = ({ typingEmulation, onUpdate }: Props) => {
         />
         <HStack>
           <NumberInput
-            label="Max delay:"
+            label="Atraso máximo:"
             data-testid="max-delay"
             defaultValue={
               typingEmulation?.maxDelay ??
@@ -79,8 +79,8 @@ export const TypingEmulationForm = ({ typingEmulation, onUpdate }: Props) => {
         </HStack>
 
         <SwitchWithLabel
-          label={'Disable on first message'}
-          moreInfoContent="When checked, typing emulation will be disabled for the first message sent by the bot."
+          label={'Desativar na primeira mensagem'}
+          moreInfoContent="Quando a opção está ativada a primeira mensagem é instantânea."
           onCheckChange={updateIsDisabledOnFirstMessage}
           initialValue={
             typingEmulation?.isDisabledOnFirstMessage ??
@@ -90,7 +90,7 @@ export const TypingEmulationForm = ({ typingEmulation, onUpdate }: Props) => {
       </SwitchWithRelatedSettings>
       <HStack>
         <NumberInput
-          label="Delay between messages:"
+          label="Atraso entre mensagens:"
           defaultValue={
             typingEmulation?.delayBetweenBubbles ??
             defaultSettings.typingEmulation.delayBetweenBubbles