+++ apps/builder/src/features/settings/settings.spec.ts
@@ -15,7 +15,7 @@ test.describe.parallel('Settings page', () => {
 
       await page.click('text="Remember user"')
 
-      await expect(page.getByPlaceholder('Type your answer...')).toHaveValue(
+      await expect(page.getByPlaceholder('Digite sua pergunta..')).toHaveValue(
         'Baptiste'
       )
       await page.click('text=Prefill input')